remove strings from ProductOpener::Lang, in order to move them to .po files
Remove all the strings that were until now stored within this module, and
move them in .po files (next commit). These files are parsed using a variant
of the Locale::Maketext::Lexicon::Gettext module, in order to keep current
hash vars unchanged.

This will certainly needs a bit more changes, to cope for example with the
path to the .po files.