Add CSS override for the fake input element.
The fake element is created by tagsinput while the original element is
hidden. This fake input element does not have any `type`, thus it was
unstyled in Foundation 5.5.1. This was changed in Foundation 5.5.3, so
that all input elements (also without a type) do get the default
Foundation styling. This CSS overrides the Foundation style, so that it
looks like the input element is not actually there.

Also added stylelint to check CSS and SCSS files for syntax problems.