Portuguese translations for the labels taxonomy
All Portuguese translations added and a few fixed, with the following exceptions:
-Added in Spanish: "Alto Contenido en Fibra" (checked in a cereals box)
-1 missing space and 1 double space in 2 french translations

At the end of document I've added labels for Portuguese protected products, not all of them but the most important (wines and meat). They are too much or there is no problem with that? Should I add all of them? (almost 100 protected products left).