crowdin suggestion
Source text should be amended to e.g.:

>> "1800+ contributors, like you, have added 13000+ products from 134 countries using the <0>Android</0>, <1>iPhone</1> or <2>Windows Phone</2> apps or cameras to scan barcodes and upload product pictures and labels.".

Obviously, non-breaking spaces (&nbsp;) *should never* be used for numbers formatting, as number formats differ in various target languages, and when this code is left out of the localization output, translators will see a completely irrelevant error message generated by the editor.
Instead, this type of separator should either be left out completely or replaced by the AE (,) or BE (.) equivalent.

I assume that the use of a space char as a numeric thousands separator is yet another French linguistic peculiarity, but this text is in English, meaning that French conventions are in no way not relevant here, sorry.