Crowdin suggestion
Many source texts in this project contain redundant words/wordings.

Unfortunately, it seems that a greater focus has been placed on ensuring a heavy use of "we", "us" & "our" (that contribute no tangible user value) in as many of the source texts as possible rather than avoiding any unnecessary redundancies such as seen in this case, which generally is a great pity.

Suggest amending this source text to e.g.:

>> "The <0>search form</0> contains dozens of criteria against which you can match products.".