Crowin suggestion
Again, probably yet another instance of erroneous (in software/related environments) use of "we", as it's assumed that no staff involvement is really necessary/required in this process/context.

Using pronouns in software (environments) doesn't meet the same goals/objectives as is the case with hardcopy material, on the contrary, in fact.