Crowdin suggestion
Since a question is asked in a sentence here, proper writing style dictates the use of an ending/concluding "?", which is currently missing.

And just to nip the potential of an error in the butt, no space char must be placed before the "?", not even a "No break space" code!