fix: Invalid facets in the url should return 404 (#8255)
if we forget any facet in the url for eg: https://world.openfoodfacts.org/category/breads/no-nutrition-data data-quality in this case, it is showing the breads page instead of an error as it is not validating whether no-nutrition-data is a valid facet or not thus generating a canonical URL and return a page. Thus we need to keep a check on this.