fix: gzip all JSON OCRs when saving OCR file on disk (#8320)
To save space, all OCR files were gzipped, but new generated files are still saved as plain text JSON files.
Also add a created_at field in the OCR JSON file containing the timestamp of generation of the OCR file. This is useful to know if we should generate again old OCR files.