deps(npm): Migrate from jvectormap to jsvectormap
jvectormap (and jvectormap-next for that matter) seem to be unmaintained.
jsvectormap is a fork, which doesn't depend on jQuery any more, and
is actually updated regularly. Also, `updateSize` doesn't seem to work in
jvectormap.

This change also moves the map initialization code from Display.pm to
it's own JS file, so that it can be linted correctly.