refactor: simplify Carrefour import for off2 migration (#9121)
Continuation of the work of to port the import scripts to off2.

On off1, we have a /home/sftp/carrefour/data dir, and we copy it to /srv/off/imports/carrefour, then we do the conversion and imports.

In practice we can do like the Equadis and Agena3000 imports and just copy the changed files from the SFTP directory to a tmp dir. The extra /srv/off/imports/carrefour is not really needed, it was just me that felt more secure with even one more copy of the data ;)

Note: I didn't notice at first that off2:/srv/off-pro/ didn't have the last pushed commits, and it had uncommited changes to the export scripts, so I commited them in this commit 31fdc36 before making my other commits on top... So I re-did some of the changes in fact. Hopefully I didn't erase anything that should have been kept.