fix: enable caching for logged-in users, use Cache-Control: no-cache (#10375)
Caching of MongoDB queries was disabled for logged in users, so that contributors could see the results of their changes immediately.

This PR changes the caching behavior so that MongoDB queries are cached for all users (even logged in users). But it enables bypassing caching when the HTTP Cache-Control: no-cache header is set (sent by browsers when doing Ctrl-Shift-R).

Also added a test.