fix: Update OTEL_EXPORTER_OTLP_ENDPOINT in .env and modperl.conf
This commit updates the `OTEL_EXPORTER_OTLP_ENDPOINT` environment variable in the `.env` file and `modperl.conf` to use the new endpoint `http://otel-collector:4318`. This change ensures that the application is sending telemetry data to the correct endpoint for OpenTelemetry export with HTTP instead of GRPC, as the Perl module doesn't understand GRPC, yet.