fix: remove non-JS list of products (#10743)
As discussed on the product opener meeting on Monday, this removes the
HTML list of products that that is displayed when browsers have turned
off Javascript. This is done for performance reasons, to avoid having
browsers download the data twice. It also makes the responses smaller
and easier to cache.