refactor: Update keycloak account link generation
This commit refactors the code in the `Display.pm` and `Keycloak.pm` files to update the generation of the keycloak account link. Previously, the `get_account_link` method in the `Keycloak.pm` file did not include the canonical URL of the current site in the generated link. This update adds the `url` parameter to the `get_account_link` method and modifies the link generation logic to include the canonical URL in the generated link.

The changes in the `Display.pm` file involve updating the usage of the `get_account_link` method to pass the `canon_url` parameter.

These changes improve the functionality and accuracy of the keycloak account link generation.

See openfoodfacts/openfoodfacts-auth#36