feat: Eco-Score improvements and change of scale: A+ to F (#10829)
The Eco-Score has an improved formula:

https://docs.score-environnemental.com/more/changelog#id-19-septembre-2024

There are 2 small changes:
- Previously products that had a non recycled and non non recyclable
packaging were capped to Eco-Score B at best. This cap has been removed.
- The country enviromental product index bonus/malus is now only applied
if there isn't already a bonus for the production system.

And there is a bigger more visible change:
- the scale is now from A+ to F, instead of A to E, in order to better
differentiate products (to avoid having most of the products of the same
category have the same Eco-Score).

Note: this is a draft. In particular, the images for the new A+ and F
grades are for testing purposes only.

To ease testing of mobile apps etc. this PR has been deployed to the
world.openfoodfacts.dev server (login and password: off)