refactor: gen_feeds_daily.pl (exports and database dumps) for all flavors (#10973)
Now the the code bases for OFF, OPF, OBF and OPFF are unified, we can
have a single gen_feeds_daily.pl file for all flavors, as the
differences are minimal (only the pro platform off-pro is different).

This PR also removes hard coded paths like /srv/off/ so that the script
can be run in a dev environment (some utilities like python, mongodbdump
etc. are missing from dev but the rest of the script can run).