fix: greenscore attribute with old ecoscore_data (#11212)
This PR fixes an issue with the new Green-Score on some products in some
countries.

When we still have the old ecoscore_data field, and not the new
environmental_score_data field, we use the older ecoscore_data to
compute attributes and knowledge panels.

However, for attributes, some parts of the code also reference
environmental_score_grade, which we didn't copy. For most countries,
this was overriden by the local scores we have in ecoscore_data. But for
countries like Russia, the bug could be seen:
https://github.com/openfoodfacts/openfoodfacts-server/issues/11184

More and more products have the new environmental_score_data field, so
the bug is less present, but it can still be seen when loading an older
revision of a product:
https://ru.openfoodfacts.org/%D0%BF%D1%80%D0%BE%D0%B4%D1%83%D0%BA%D1%82/8076800195057/spaghetti-n-5-barilla-norge-as?rev=3517

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>