feat: taxonomy enhancer (#11267)
### What
Based on the following comment on Slack:
> There is a single language used to parse ingredients in Open Food
Facts (based on ingredient_text, and leading to ingredients,
ingredients_lc, ingredients_analysis, and more key values in the api)
> All languages for a product are given in ingredients_text_<language>
> That might already have been discussed but would that be an option to
parse ingredients of all languages?
> Eventually, given some conditions on ingredients_text_<language> (no
quality errors/warning like @ symbol for example)
> It would make us slowly go toward an autocompletion of ingredients
taxonomy. Example: if for the same product, you have 3 ingredients in
English, all known. And 3 ingredients in Greek, only 2 known. And the
position of these 2 corresponds to the position of their English
translation. Then, we could safely add the entry of the unknown Greek
ingredient to the ingredients taxonomy. With a PR, so that it can be
reviewed.
> We could start with very strict conditions like for this example no
more than 3 ingredients + only one unrecognized ingredient.
> We could also create quality info if text is way too long in a given
language whereas all other languages have a shorter ingredients list.
That would help to find stop words. Instead of waiting for contributors
in each language.
> It would allow to improve data quality as well.


The present PR aims to tackle the following:
- detect stop words before ingredients list (missing_stop_words_before),
and stop there if it finds something for that language, otherwise
- detect stop words after ingredents list (missing_stop_words_after),
and stop there if it finds something for that language, otherwise
- compare ingredients one by one to:
  - detect missing translation in the taxonomy (missing_ingredients)
- typo in ingredients list (could have been a dq error but->) or typo in
the taxonomy (ingredients_typo)
- ingredient having different ids between languages and those ids do not
have a parent/child relationship (mismatch_in_taxonomy). This latter is
ignored if it leads to more than 1 occurence because it can be due to
new version of the product ingredients list and one language having new
version while another language is having old version, in which case
there is overlap between ingredients id but if order of ingredients is
not the same as before it will find many mismatches, see below
8014190017627 example.

**Integration**
For devs purpose, it is using some facets, similar to data quality
facets. Adding a route like this
https://hr.openfoodfacts.org/data-quality-errors would make it easy to
review these new facets.

These new facets cannot be data quality facets because a contributor
cannot fix things by editing products, one needs to modify the
taxonomies files or the code of ProductOpener.



### Screenshot

![Screenshot_20250119_100623](https://github.com/user-attachments/assets/6ae0bb80-fb1c-4778-b78a-288f60247ef8)



### Related issue(s) and discussion
- Fixes #-none-

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>
Co-authored-by: Stéphane Gigandet <stephane@openfoodfacts.org>