taxonomy: Copy some `se` entries to `sv`
`se` is the language code for Northern Sámi, but all of the copied
entries are in Swedish (`sv`). It is common to mistakenly use `se` for
Swedish, since `se` is the country code for Sweden.

I didn’t remove the Sámi entries since North Sámi _is_ spoken in parts
currently occupied by Sweden, so there is likely some loan words and
other linguistic influences. I have my doubts with a lot of these
though, but I’m not a Sámi speaker of any sort and I don’t know the
circumstances of why they were added.