fix: data quality label threshold source of vitamins and minerals (#11536)
### What
There is a bug with quality facets for this product:
https://world.openfoodfacts.org/product/5411188128311/coconut-unsweetened-alpro

Calcium source label claim but calcium below 0.8
Vitamin b12 source label claim but vitamin b12 below 1e-06
Vitamin D source label claim but Vitamin D below 5e-06

Here is the reference EU regulation document:
https://eur-lex.europa.eu/legal-content/EN/ALL/?uri=CELEX:31990L0496
(see Annex)
See also
https://eur-lex.europa.eu/legal-content/EN/TXT/?uri=CELEX%3A02006R1924-20141213
for difference between source of something and rich in something

The threshold for all vitamins and minerals in the reference EU
regulation document should not be the values given but 15% of the
values.

\+ also update thresholds values for jam, extra jam and jelly to reflect
last version of EU regulation.


### Related issue(s) and discussion
- Fixes #11531