fix: TaxonomiesEnhancer division by zero (#11607)
We had an issue to publish products from the Systeme U import, it took
quite a bit of time to debug, so I wrote a little doc of what I did
(maybe not the best way to do it). It turns out to be a small issue in
the logging statements of TaxonomyEnhancer.pm.

In Perl $#array is the indice of the last member of the array, which
starts at 0. So if the array has 1 element, its value is 0. So on a
specific product we got a division by zero.

minion_log.conf has logging at the debug level, so it may be a reason
why we got this error in the minion tasks and that we didn't get
division by zero (or at least we didn't see them) in the website / API.

---------

Co-authored-by: Alex Garel <alex@openfoodfacts.org>