feat: taxonomize brands with a language less xx: taxonomy (#11606)
This PR is intended to replace
https://github.com/openfoodfacts/openfoodfacts-server/pull/10662 which
got broken with the merges.

This PR is a proof of concept to make brands a taxonomized field using a
language less taxonomy (all entries use the xx: prefix).

To taxonomize existing products:

~/openfoodfacts-server$ docker exec -it po_off-backend-1 bash
www-data@a01d17ec9489:/opt/product-opener$
./scripts/update_all_products.pl --fields brands
Fixes https://github.com/openfoodfacts/openfoodfacts-server/issues/5208

This PR is a proof of concept, it lacks tests, documentation etc.

Also the "brands" taxonomy is hardcoded in Tags.pm, it would probably be
better to make this configurable in case we want to do something similar
with other taxonomies in the future.

Read API requests should still get the same values in the "brands" field
(but there will be new fields brands_tags and brands_hierarchy).

Write API requests should continue to use the "brands" field to write
brands.