refactor: fix for ingredients count + schema version / API version conversion (#11615)
This PR fixes #11610 by simplifying how we count ingredients and known
ingredients.

It removes from products the array ingredients_hierarchy which content
is identical to ingredients_tags

As we don't want to break older clients that may use
ingredients_hierarchy, we will put it back when the API version is less
than 3.2

This PR also introduces a new system to manage schema changes:
- we will add a schema version to each product object
- products schema will be converted as needed (upgraded when reading
older product revisions, and downgraded when the requested API version
is lower than the current version).

We can also add documentation about the schema changes and the
corresponding API version changes.

---------

Co-authored-by: Alex Garel <alex@openfoodfacts.org>
Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>