feat: add diet info from GS1 imports + match categories with gpc_category_code:en (#11783)
GS1 has ~ 20 diets (e.g. vegan, vegetarian etc.) that product can be
tagged with. This PR adds support to add those diets as labels.

Also included: we now match categories using the gpc_category_code:en
property in the categories taxonomy. Previously we only did a match on
the GPC category name, and not the code.

+ some tests

---------

Co-authored-by: Alex Garel <alex@openfoodfacts.org>