feat: Add a button to add product if not found in search (#11567)
### What

This PR adds a "Add" button in '/cgi/search.pl' if no product was found.
Also added "Add it to our database, it only takes a few minutes :)" text
to encourage people to update the database.

This
[issue](https://github.com/openfoodfacts/openfoodfacts-server/issues/9173)
mentioned to add a button to create new product (if not present in
database) for the barcode searched by user.

But with this PR, button shows up even if the searched item was not a
barcode. Should I make modification to show button only for barcode. In
that case, is there a function to find out whether a text is valid
barcode or not. Or if function isn't present then please tell me what is
considered valid barcode.


### Screenshot

![{4006C9F3-CA6C-43F0-B638-1A094FD0EB1B}](https://github.com/user-attachments/assets/04de491c-4007-4276-a710-979970ff9a2e)

**In edit mode**

![{F970849A-8CB2-4BAE-BA71-9AE1AF5EF553}](https://github.com/user-attachments/assets/2b2c0a0e-326f-448c-8a03-ec12bd6ec3a5)



### Related issue(s) and discussion

- Fixes #9173

---------

Co-authored-by: Nothingspecial99 <vikaspandey.vp.ap@gmail.com>
Co-authored-by: Alex Garel <alex@openfoodfacts.org>
Co-authored-by: Jagjeevan <83103469+JagjeevanAK@users.noreply.github.com>