fix: use 'yes' instead of 'en:yes' for vegan / vegetarian ingredients properties (#11829)
In the ingredients structure, we have inconsistent values for the
"vegetarian" and "vegan" properties. In the taxonomy, they are set to
"yes", "no", "maybe", but in the ingredients analysis, if there was a
label or vegetal origin (e.g. "coagulating enzyme (vegan)", we would put
"en:yes" as a value.

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>