feat: Handle label/ingredient conflicts (#11831)
### What

Sometimes a product may contain ingredients that directly contradict a
given label. E.g., a “vegan” product might contain milk. This can be due
to a variety of factors and might just be bad data (we already have data
quality checks like “vegan label but not vegan ingredient”), but might
also be a genuine error on the product or producer’s side.

Right now, a label will always override whatever is discovered during
ingredient parsing, even if they directly contradict each other. E.g., a
product with “palm oil” as an ingredient and “no palm oil” label would
be given as `en:palm-oil-free` which may (or may not!) be providing
false information to users/consumers.

Instead of outright claiming that a product is vegan, vegetarian, or
palm oil free because there’s a label on the packaging, this checks the
ingredients to see if anything contradicts this statement. If so, this
now demotes the product to a “maybe” status to let the user/consumer
know that there is something going on and exercise extra caution.

### Related issue(s) and discussion
- Slack discussion:
https://openfoodfacts.slack.com/archives/C03H290LF/p1743584095493549
- Fixes
https://github.com/openfoodfacts/openfoodfacts-server/issues/11830