taxonomy: Tighten up vegan requirements
A number of ingredients that are on their surface vegan actually undergo
some kind of treatment or processing using animal products, which would
make them not vegan. In many cases, this would get reflected in
ingredient lists, like gelatin in jams and candy etc., but in some cases
this is done in ways that are not going to be reflected in an ingredient
list, e.g., because there won’t be one—like for fruits.

This also means that just because something is “vegetable”, “extracted
from plants”, “plant based”, etc. it doesn’t automatically make them
vegan, so those labels have been moved to be synonyms (or children) of
vegetarian rather than vegan.

This approach is the same that is already used in `beauty/labels.txt`
where “no animal derived ingredients” implying vegetarian status rather
than vegan status:
```
< en:Vegetarian
en: No animal derived ingredients
```

Additionally, a number of labels have been set to be children of
`en:Non-vegan`, either directly or indirectly. With the label overriding
happening when analysing products, this should also help to catch more
non-vegan products that may currently get set to unknown or maybe status
because of unknown ingredients.
(This includes the label `en:Non-vegetarian`. If something is not
vegetarian, it won’t be vegan either.)

Some other, smaller changes to catch more non-vegan things:

Ingredient `en:animal` is set to non-vegan, as anything that is derived
from a non-human animal is, by definition, not vegan.

Vitamins have been changed from `vegan:en: yes` to `maybe`, as some
vitamins may be extracted or otherwise derived from non-human animals.

Some ingredients have had `vegan:en:` and/or `vegetarian:en:` set, as
they weren’t recognised as either via parents or otherwise.

A new `en: Vegetarian products` category was added and some
categories were set to be `incompatible_with:` that or the
`en:Vegan products` category.