taxonomy: Add aluminium foil as a material
Right now aluminium foil is covered by the `packaging_materials`
`en:aluminium` combined with the `packaging_shapes` `en:foil` (which is
a synonym for `en:film`).

This misses out on some points though, primarily that there are a lot of
translations and synonyms for aluminium foil that do not lend themselves
to this simple combinations.

E.g., `en:tin foil` and `da:tinfolie` refers to a different material
(namely tin) which should _not_ be a synonym for aluminium.
(`da:stanniol` is also a name derived from “tin”.)

Another example is `da:sølvpapir` and `se:silbabábir` which translates
to “silver paper”. Again: a completely different metal/material, as well
as a shape that is also a material.

Adding “aluminium foil” as its own, discrete packaging material allows
us to deal with these translations and synonyms, and also allows us to
attach additional metadata (such as `wikidata:`) to it.