feat: identify products with rare crops - DIVINFOOD NUCs (#11913)
This PR adds a misc tag to identify products with ingredients that
containt rare crops / neglected and underutilized crops (NUC). It is
needed for the DIVINFOOD EU project.

Also some new Hungarian translations for rare crops, and a tool to
generate a table of taxonomy translations, so that it's easy for the
DIVINFOOD partners to review the translations.