fix: remove old unused / debug fields (#11930)
This PR removes a bunch of old unused / debug fields. Some of them were
already not generated anymore but can still be found in old products.

It also removes some code that is not used anymore:

- the ability to load multiple taxonomies for the same tag type ("prev",
"next" and "debug") to list products that have different tags for a new
version of the taxonomy: while interesting, in practice it has not been
used for the last 5 years at least
- extended ecoscore data: was implemented as an experiment in 2020, but
was never activated in production

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>