taxonomy: “Orphan” Kött/Mjölk från Sverige
Kött and Mjölk från Sverige specifically refers to meat and milk
respectively and doesn’t necessarily apply to the rest of the
ingredients (I think?). Even if it does, if there’s only the Kött or
Mjölk label on packaging and no “Från Sverige”, we probably don’t want
to show both logos on product pages, which would be another result of
this inheritance.