taxonomy: move processing from ingredients (#11811)
### What

#### Remarks:

- **Remark 1**: lost of ciqual

beside the benefit of 1) reducing the taxonomy file size and 2) having
visibility of processing methods, the ciqual codes - for milk powder for
example - are removed from the taxonomy.

- **Remark 2**: some ingredients have been added in the past (by me) to
cover percentage of fat in the milk. Percentage can also be moved in the
processing.

Example:

> en: skimmed milk with 0.1% lactose-free milk fat
> hr: obrano mlijeko s 0.1% mliječne masti bez laktoze
> it: latte scremato con 0.1% di grassi senza lattosio

It leads to duplicates in the processing key-value for the ingredient. 

> "processing": "en:lactose-free,en:skimmed,en:skimmed",


Duplicates could be removed maybe in the future?

- **Remark 3**: in some languages a single word defines skimmed-milk and
semi-skimmed-milk. This is problematic.

> et: Lõss, lõssi, lössi
> is: Undanrenna, undanrennu
> pa: ਸਪਰੇਟਾ

- **Remark 4**: for some ingredients the processing method is in the
middle of the inredients

> de: Kuhrohmilch


But for most of the cases, the processing is extracted beside being in
the middle of the ingredient. There might be already something in place
to tackle that in the code but not covering all cases

- **Remark 5**: added some languages to parse the processing from the
ingredients in Ingredient.pm, but also found some exception that need to
be tackled

For Example, 'hu' already "match after the ingredient, does not require
a space" "match before the ingredient, require a space" but also should
match before without space, like for tejpor





### Related issue(s) and discussion
<!-- Please add the issue number this issue will close, that way, once
your pull request is merged, the issue will be closed as well -->
- Fixes #[ISSUE NUMBER]

---------

Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>