feat: Implement OIDC sign-in (#9251)
### What

The main aim is to use Keycloak as an IdP, and connect Product Opener as
a RP. We want to move the basics of user management out of Product
Opener.

In the future, this can be used as a base to enable #1204 to be
implemented, even though it will probably not be part of the initial
change as to not blow up the PR too much. One important requirement is
to keep compatibility with the current Basic Auth mechanism.

### Screenshot

<!-- Optional, you can delete if not relevant -->
TBD

### Related issue(s) and discussion

<!-- Please add the issue number this issue will close, that way, once
your pull request is merged, the issue will be closed as well -->
- closes #1596

Adding a task list to track progress:

- [x] @hangy: Move all interaction with user_emails.sto to call Keycloak
APIs instead
- [x] @john-gom: Move all read and write operations to user.sto files to
one place
- [x] Update user.sto read/write operations to get user "signup" fields
from Keycloak (not required, as we want Keycloak as the canonical
source)
- [x] @hangy: Pass country / language to login page
- [x] @hangy Add delete user event listener to Keycloak, trigger Redis
event openfoodfacts/openfoodfacts-auth#3
- [x] @hangy Store deleted users in a table
openfoodfacts/openfoodfacts-auth#14
- [x] @hangy Redis event subscriber on Product Opener to remove user
name from products (TBA extend to other applications) (WIP
https://github.com/hangy/openfoodfacts-server/tree/keycloak-redis)
- [x] @john-gom: Add all "signup" fields to Keycloak and import language
and country lists from taxonomy json
- [x] @john-gom: Fix locale pick list to be searchable
- [x] Upgrade to 24.0.0 when available and check localization is working
for account page
- [x] Set up Crowdin translation project
openfoodfacts/openfoodfacts-auth#2
- [x] @hangy Test with Mobile and other API usage scenarios
  - Flutter app seems to work fine
- Hunger games shows user logged in (with modified hunger games branch,
as `main` currently has a lot of hard coded URLs)
  - Write test with robotoff not that easy in separate setup 😞 
- [x] @hangy Benchmark import of the number of users we have in
production
`time ./scripts/migrate_users_to_keycloak.pl` with around 400.000
synthetic accounts took about 45 minutes on my desktop PC in WSL2.

Save for a later phase:

- [ ] HOLD on this pending unification. Live with inconsistent passwords
for now. Update OPF, OBF and OPFF to login via Keycloak and redirect to
OFF for user account info
- [ ] Extend user attribute table to store more text
- [ ] Select correct language / country for user after login redirect
back to OFF
Example: If the users was on `fr.openfoodfacts.org` before signing up,
and they change their locale to Spanish during registration in Keycloak,
do we want them to be redirected to a Spanish page instead?
Problem: Differentiating between first login (because of registration)
and subsequent login would have to be done based on whether or not
`${userid}.sto` exists.
- [ ] Get Product Opener client to self register on initialization and
store secret somewhere safe
Advantage: The client secret does not have to be exposed to the admin.
However, they'll have access to the secret anyways, unless we were to
store the secret in some kind of HSM.
Disadvantage: Some client configuration can be done during
self-registration, but some necessary permissions like realm-management
need to be configured manually by an admin, anyways.
- [ ] Get rid of user.sto files completely
To consider: This file currently contains information about the users'
active sessions. Do we want to replace the opaque session identifier by
an encrypted cookie? Also, we need some kind of directory to identify
users that have used ProductOpener know if we have to do something when
a deletion event comes in through Redis. (Obviously, this could be done
in MongoDB or PostgreSQL instead, but is there a huge difference?)
- [ ] Move API consumers to use client_credentials rather than login
- [ ] Update mobile app to use Keycloak web forms for login (eg. OIDC
Authorization Code Flow)
- [x] ~~Figure out how/if org management should be done in Keycloak~~
(Done in MongoDB)

Moved to `openfoodfacts-auth`:
- [x] Create message files for all fields we will need translating
openfoodfacts/openfoodfacts-auth#33
- [x] Finalize theming for all Keycloak screens that users will see
openfoodfacts/openfoodfacts-auth#31
- [x] Configure signup / forgot password email in Keycloak
openfoodfacts/openfoodfacts-auth#32
- [x] figure out security configuration for production deployment
openfoodfacts/openfoodfacts-auth#34
- [x] work out what to do about staging, e.g. share keycloak with
production or have separate instances and maybe copy database across
periodically openfoodfacts/openfoodfacts-auth#35
- [x] Support social login providers openfoodfacts/openfoodfacts-auth#49
- [x] Support passkeys openfoodfacts/openfoodfacts-auth#50

---------

Signed-off-by: John Gomersall <thegoms@gmail.com>
Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: John Gomersall <thegoms@gmail.com>
Co-authored-by: john-gom <116556069+john-gom@users.noreply.github.com>
Co-authored-by: Open Food Facts Bot <contact@openfoodfacts.org>
Co-authored-by: Raphaël Bournhonesque <raphael@bournhonesque.eu>
Co-authored-by: Alex Garel <alex@garel.org>
Co-authored-by: Alex Garel <alex@openfoodfacts.org>
Co-authored-by: James Addison <55152140+jayaddison@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Frederik “Freso” S. Olesen <freso.dk@gmail.com>
Co-authored-by: Stéphane Gigandet <stephane@openfoodfacts.org>
Co-authored-by: Open Food Facts Bot <119524357+openfoodfacts-bot@users.noreply.github.com>
Co-authored-by: Kout95 <65901733+Kout95@users.noreply.github.com>
Co-authored-by: Charles Nepote <charles@nepote.org>